Fetch metrics from multiple adaptors
Adaptor updates and metric updates are separated by a lock.  We may have
to wait a short bit to update the metrics while an adaptor's information
is being filled in, but that will be much less time than waiting for an
adaptor to be read (which might timeout or fail).